requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.3.2
Pillow==10.2.0
supabase==2.3.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
